refactor: move router setup back to main.tsx and remove router.ts
- Move TanStack Router configuration back to main.tsx
- Remove utils/router.ts as the setup is simple enough for main.tsx
- Keep only utils/queryClient.ts for complex query configuration
- Simplify utils structure while maintaining clean separation
